Split client.py into a client package
